soc: octeontx2-sdp: Fix SDP output backpressure
commit 8d5980831a1d1a8d15323fb2ed6da9cc0b987798 from
git@git.assembla.com:cavium/WindRiver.linux.git

The backpressure only works when a single SDP queue per VF or PF is used. This
is due to how SDP queues interract with NIX output TL4 queues, and the
requirement that there must be a 1-1 TL4 to SDP queue mapping for backpressure
to work.  In order for backpressure to work with more SDP queues, NIX queue
allocation for SDP must be changed to match.

Signed-off-by: Roy Franz <rfranz@marvell.com>
Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: Ia2a7adbc9dad8a7f034985edb884a92061eaa1ee
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>