octeontx2-pf: Limit number of QOS queues
commit fd06ad3fded99a6f577d54c6894455543c577c6d from
git@git.assembla.com:cavium/WindRiver.linux.git

Receive/trasmit irqs and cpus are one to one mapped for better
performace. When number of cpus are lesser than Qos queues the
msi vector allocation is failing to fix the issue  number of
Qos queues are limited to number of cpus.

Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Change-Id: Icb81c09ed8e610997406abf32ead73b2dca6a4f5
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/96841
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>