octeontx2-af: Update correct mask to filter IPv4 fragments
commit cccd24be672056473b1d33623b3378285ab52dcd from
git@git.assembla.com:cavium/WindRiver.linux.git

During the initial design, the IPv4 ip_flag mask was set to 0xff.
Which results to filter only fragmets with (fragment_offset == 0).
As part of the fix, updated the mask to 0x20 to filter all the
fragmented packets irrespective of the fragment_offset value.

Fixes: a35d7094a429 ("octeontx2-af: Add support to filter packet based on IP fragment")
Signed-off-by: Suman Ghosh <sumang@marvell.com>
Change-Id: I5c11fd62d2090ad777d776c4c0103039a01ef4c7
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/97383
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>