MLK-25315: dma: pxp: add monochrome mode support for PS engine
commit  14c658fa7661843e68929243443e24717a7cd52d from
https://github.com/nxp-imx/linux-imx lf-6.12.y

When input format of PS engine is monochrome, such as Y8 or Y4,
the low 16 bits of PS_VBUF are used as the U/V data in the data
path instead of sourcing U/V data from external buffers. In this
case it represents a fixed value for U/V data.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>