MA-21077 mfd: Remove __init from fp9931_exit
commit  b7594bb8d8cd2994dcd2ae8f0e73153cb8486d15 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The function fp9931_exit is defined with the __init annotation, which
means that it can only be called during the initialization phase of
the module. However, this function is registered as the
module exit function using the module_exit macro.

To fix this issue, remove the __init annotation from the
fp9931_exit function.

Change-Id: I7bc17298577c44a9a67445be4a479234f7298a36
Signed-off-by: Zhipeng Wang <zhipeng.wang_1@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>