LF-8475 arm64: dts: imx8mq-wevk: enable the front M.2 connector on imx8mq-wevk board
commit  f327e3d202d2548003482440642c93b0d8cb71bd from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The main change for imx8mq-wevk board is to deprecate the on-board
cypress wifi and add a new M.2 on the front of the board. This new
imx8mq-wevk.dts enables the front M.2 connector for WiFi and Bluetooth.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>