phy: lynx-28g: add support for 25GBASER
commit  5cb5eeb5dd9856f2fef2ebda8432f1857cf564cc from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Add support for 25GBASE-R in the Lynx 28G SerDes PHY driver.
This mainly means being able to determine if a PLL is able to support
the new interface type, to determine at probe time if a lane is
configured from the Reset Configuration Word (RCW) with this interface
type and to be able to reconfigure a lane.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>