arch: arm64: lx2160a-rdb: describe the retimer as a Generic PHY
commit  d0a692e3b8dd125d4f7f609ff9d7162caf17eedc from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The LX2160ARDB board has a Inphi Retimer PHY which is used on the 2
SerDes which go to the SFPs. Model this Retimer PHY as a generic PHY and
extend the MAC DT nodes to also reference the retimer.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>