gpio: regmap: add the .get_direction() callback
commit  2fb972506888b1f4916ebd5f769566243f1d2f05 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

There are GPIO controllers such as the one present in the LX2160ARDB
QIXIS CPLD which are single register fixed-direction. This cannot be
modeled using the gpio-regmap as-is since there is no way to
present the true direction of a GPIO line.

In order to make this use case possible, add a new callback to the
gpio_config structure - .get_direction() - which can be used by user
drivers to provide the fixed direction per GPIO line.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>