rpmsg: imx_rpmsg: Change the notifying to timeout mode
commit  92f6ed1bd738bd07632856f0ea761627e9eefc28 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Currently, after the rpmsg link up it will use mailbox
channel in blocking mode to send the notifications. So
it will be stuck in the notification when the remote side
unable to respond for some reason, such as crashed.

To avoid interlock, this patch changes the notifying to
timeout mode, and treats the virtqueue as broken when
the remote side doesn't respond within 1 second.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>