LF-9849-4: net: phy: tja11xx: do not enable both REFCLK_IN and OUT when initializing
commit  477ce435a09668db045c7acb2187fa7033611c89 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The BMCR_PDOWN bit is set when initializing the tja1100 configuration
to avoid conflicts if both REFCLK_IN and REFCLK_OUT are enabled. In this
way, it can avoid the issue of the configuration unable to take effect
even after the registers are correctly programmed.

Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>