net: fman_memac: refactor pcsphy-handle acquisition
commit  ed9b625b986d0eed8d1ab37313f3e4dd038a265d from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The way in which pcs-handle-names is parsed is a bit repetitive, and the
detection of the default case (no pcs-handle-names) is confusing.
The code assumes that no pcs-handle-names array exists if there is no
"sgmii" entry in it, which is not quite correct. It is perfectly legal
to have pcs-handle-names without "sgmii".

Create memac_get_optional_pcs() for looking up based on
pcs-handle-names, and memac_get_default_pcs() for when that lookup could
not take place.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>