LF-10639-4 net: pcs: xpcs: add mx95 serdes support
commit  61ddb1408f5b65d1fb6bfa367c12bd78f61cbf00 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The serdes and pcs in the mx95 chip are closely related.
Both serdes and pcs are accessed through the IMDIO bus.
PCS's clock is provided by serdes' clock. Before accessing the ID register
of PCS, need to lock the access permission by accessing the MAC_ADAPTER
register of serdes, and then check whether the PCS is initialized normally.

This is different from the initialization sequence in the original PCS driver.
Therefore, add xpcs_create_mdiodev_with_phy() function to initialize xpcs
with serdes.

Because the ID of mx95 PCS is the same as the default ID of XPCS,
the initialization function for pma/serdes cannot be added.
In mx95, use the PMA's ID to match xpcs_id.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>