LF-10932-2 nvmem: imx95: add the support to return the mac addr of PFs
commit  74efea1cd0b75dded01e08113517bb6155713e40 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The 3 PFs of imx95 use 9 mac addresses, but only saves one mac base
address in fuse, which is the mac address of the first PF. The others
PF mac addresses need to be calculated through offsets.
According to the ocotp framework, adding the PF offset to bits 15~12 of
offset. This will not affect the dts node setting of other fuse devices.
In imx95, the value range of offset is 0~0x984.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Alice Guo <alice.guo@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>