LF-11414-6 media/i2c: add max96717 chip library
commit  ab109353b9147b01b3699f991781fe7b1daea702 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The MAX96717F GMSL serializer receives video on a MIPI CSI-2 interface
and outputs it on a GMSL2 serial link transceiver.

This is just a library that will be used by camera modules drivers that
usually contain a serializer and a camera sensor.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Mirela Rabulea <Mirela.rabulea@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>