LF-11280 drm/bridge: imx95-pixel-link: Fix output endpoint
commit  d70c22628439d4ba55ce5eaf0eba56e3aaf5b6c6 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The endpoint parameter of of_graph_get_remote_node() is a 32bit unsigned
integer.  imx95_pl_find_next_bridge_per_output_port() calls the function
in a 'for' loop to get remote node starting from endpoint1 and uses
decrement operator to decrease the endpoint one by one.  That underflows
the endpoint as an integer and thus the 'for' loop check is always true.
Fix the endpoint by changing the way to calculate it out so that the
endpoint is either 1 or 0.

Signed-off-by: Liu Ying <victor.liu@nxp.com>
Reviewed-by: Sandor Yu <Sandor.yu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>