dpaa2-switch: add dpaa2_switch_port_to_bridge_port() helper
commit  7c2f47548fc8e282ef4e4ff5fc98aa189a13dd8b from
https://github.com/nxp-imx/linux-imx lf-6.12.y

With the addition of offloading support for upper bond devices we have
to let the switchdev framework know if a specific bridge port is
offloaded or not, even if that port is bond device.

For this to happen, create the dpaa2_switch_port_to_bridge_port function
which will determine the bridge port corresponding to a particulat DPAA2
switch interface and use it in the switchdev_bridge_port_offload call.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>