net: tsn: remove unused notifier chain functionality
commit  95034e2c4a80f7fcaae3b42a1a5114a77d957d06 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Apparently it was intended for drivers to emit TSN_QBV_CONFIGCHANGETIME_ARRIVE
and for this to notify user space, but it seems to have never been
implemented.

If it ever is, it is rather likely we will want to expose this through
tc instead, so just drop the unused code from the legacy tsntool.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>