soc: fsl: guts: use a macro to encode the DCFG CCSR space
commit  9d4d69e7700e9b4ccd4745d19c044a63eec90b94 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Instead of using a hardcoded value when iomapping the DCFG CCSR space,
add a new macro for it. The code will be easier to follow this way,
especially when we add support for the DCFG DCSR space as well.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>