arm64: dts: ls1088ardb: describe the on-board TI retimer
commit  42d0343f2cd25b9f1cd0f071cde20d5e46867683 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Now that we have a driver for the on-board TI retimer, we can fully
describe it in DT and take advantage of it by referencing it from the
ethernet node.

With this patch in place, we are now able to dynamically reconfigure
(only at boot time, since no MOD_ABS GPIO connected to the cage) the
DPMAC.1 ethernet interface depending on the SFP module inserted.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>