LF-11561-1 arm64: dts: imx8qxp-sof-cs42888: Use fsl,imx8qxp-mek-bb compatbile
commit  75e701874b852667df3aed89de94db4fbacae462 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Ideally, we wouldn't care if the baseboard is installed or not
and ideally we would create a separate sound card for each codec
we support.

Unfortunately, SOF doesn't support multiple sound cards as of today
and we have separate dts for usecases when we enable wm8960 (found on
cpu board) and cs42888 codec found on Audio IO card which is connected
to CPU board via baseboard (BB).

As an workaround we override the board compatible so that we know we
are using the baseboard, and thus we know that we need to select proper
topology to use cs42888 codec.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>