LF-11321-3 arm64: dts: imx95: add new dts for NETC RPMSG use case
commit  bc475edff4a261dcf3c1f03647a6796306598475 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Add new DTS for NETC RPMSG in which case only the two VFs of ENETC2 are
owned by A core, other functions are all owned by M core, including the
entire configuration space of NETC is owned by M core. That is to say, A
core cannot directly access to the configuration space, so we allocate a
shared memory as virtual configuration space to store the information
from the real configuration space. So that A core can directly obtain
the information from the shared memory. In addition, M core will update
the shared memory according to the relevant operations. The memory range
is from 0x88224000 to 0x88323FFF.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Reviewed-by: Frank Li <frank.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>