LF-12044-1 arm64: dts: imx93: remove pinctrl-assert-gpios property for WiFi&BT
commit  33e877d679f406ab61abef6e0e0952adb7ac0782 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

For imx93 11x11-evk and 9x9-qsb board:
The pinctrl-assert-gpios property in usdhc3 dts node is used to control
the M.2 power, we should use the reg_m2_pwr regulator and set it as the
parent regulator of reg_usdhc3_vmmc.
The pinctrl-assert-gpios property in lpuart5 dts node is used to control
the M.2 PCM pins level shifter, we can use the gpio-hog to set the
defalut gpio status to replace pinctrl-assert-gpios.

For imx93-14x14-evk board, pin 19 of pcal6524 is not used, so no need to
configure it, remove it here.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>