LF-12040-1 net: enetc: adjust the management of NETC clocks
commit  8cd252b7225452b05d778e382f232bda2a725206 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The netc_prb_ierb driver is probed earlier than other drivers of
NETC, such as enetc, emdio and timer. So it's better to move the
system clock to netc_prb_ierb driver.
In addition, it's more reasonable to enable the reference clock
when opening the interface and disable the reference clock when
stopping the interface.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>