PCI: dwc: Fix resume failure if no EP is connected at some platforms.
commit  85ec9c75507e4c62e16d49b421f63ca1d4489f85 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The dw_pcie_suspend_noirq() function currently returns success directly
if no endpoint (EP) device is connected. However, on some platforms, power
loss occurs during suspend, causing dw_resume() to do nothing in this case.
This results in a system halt because the DWC controller is not initialized
after power-on during resume.

Change call to deinit() in suspend and init() at resume regardless of
whether there are EP device connections or not. It is not harmful to
perform deinit() and init() again for the no power-off case, and it keeps
the code simple and consistent in logic.

Fixes: 4774faf854f5 ("PCI: dwc: Implement generic suspend/resume functionality")
Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>