driver: sdk_fman: move the irq free operation to the front of releasing register resource
Based on current code, if CONFIG_DEBUG_SHIRQ is enabled, there is below
calltrace when executing reboot operation.
[  225.626580] Unable to handle kernel paging request at virtual address 006b6b6b6b6b6c3f
......
[  225.797449] Call trace:
[  225.799890]  fman_get_normal_pending+0x14/0x34
[  225.804333]  __free_irq+0x234/0x334
[  225.807821]  free_irq+0x3c/0x8c
[  225.810959]  devm_free_irq+0x64/0xa0
[  225.814534]  FreeFmDev.part.0+0x50/0xc4
[  225.818369]  fm_shutdown+0x30/0x4c
[  225.821769]  platform_shutdown+0x2c/0x3c
[  225.825691]  device_shutdown+0x130/0x210
......
[  225.860374]  el0t_64_sync+0x18c/0x190
The root cause is that the fm_irq() is invoked as a callback function,
and there is register operation in fman_get_normal_pending() after register
resource has been unmapped. In this way, above calltrace is triggered and
system hangs up. Therefore, move the irq free operation forward to make sure
that it is before releasing register resource.

Signed-off-by: Meng Li <Meng.Li@windriver.com>