LF-12998-1 net: enetc: populate NETC iEPRC nodes from netc_prb_ierb
commit  a516344dc7ce5421e368897bc011a420ffd35253 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

As the IERB is used for pre-boot initialization, so all the devices
of NETC depend on IERB module, such as LDID and memory configuration
and so on. So IERB must be initialized before the initialization of
any devices of NETC.

In the previous implementation, we used netc_ierb_get_init_status()
to ensure that enetc, emdio and timer drivers were probed after IERB
completed initialization. However, it can only solve the loading order
issue of IERB and other NETC device drivers. But it cannot solve the
execution order of PM interfaces of netc_prb_ierb and other drivers
when doing system suspend/resume. Therefore, in order to highlight
the dependency relationship between IERB and other NETC devices, we
make the NETC iEPRC modes as children nodes of netc_prb_ierb node.

In addition, since the emdio driver is the supplier of the mdio bus
and the enetc driver is the consumer, this binding relationship needs
to be established to avoid the issue of accessing NULL pointer when
the emdio driver is unbound. Therefore, the enetc driver must be
unbound before the emdio driver. Also, the enetc driver must be probed
after the emdio driver completes the probe.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>