MMFMWK-9397: arm64: wave6: remove level interface
commit  d85f1307f42ebf37f6a6ef75761e00988c307072 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

For 4k decoding in our board, the performance can reach higher level than
which marked in datasheet.
So removing level interface of decoder is reasonable.
Otherwise, We can not ensure that the encoder will not use a large level
when encoding with a small bitrate or resolution.

Signed-off-by: TaoJiang <tao.jiang_2@nxp.com>
Reviewed-by: MingQian <ming.qian@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>