LF-12807-6 net: enetc: don't reconfigure if XDP is already enabled
commit  64655404ff49f714ab83be37efe6bdf13e36bacf from
https://github.com/nxp-imx/linux-imx lf-6.12.y

In the previous enetc driver, the Tx and Rx BD rings will be reconfigured
when a bpf program is installed. But it is not necessary if the XDP is
already enabled. So add an optimization for enetc_setup_xdp_prog(), the
BD rings won't be reconfigured and only update bpf program if XDP is
enabled before.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>