dpaa2-eth: add support to query number of Tx Tcs
commit  7e78c90a7a26aa100aaedb8b1cebe68c340aa8a9 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

Newer versions of the dpni_get_attributes() MC API will return the
number of Tx TCs separately from the Rx once. This is especially useful
for the LX2160A SoC usecase which supports max 16 Tx TCs, while only 8
Rx TCs.

Extend the dpaa2-eth driver so that it can query the number of Tx TCs
for the current DPNI. Since we do not want to break the usecases in
which an older MC version is used, the newer version of the command is
used only after the DPNI version is compared to the minimum required
one.

The dpaa2_eth_tc_count() macro is also split into two new macros, one
for Tx or for Rx TCs, and its usages replaced by the new ones.
The call to the dpaa2_eth_detect_features() function is moved much
sooner in the probe path of the device so that we can use the feature
bitmap earlier.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>