net: pcs: lynx: warn if SerDes MDEV_PORT differs from our mdiodev->addr
commit  be42e6d1afc6ecdf37166f47b2832b5c28dfd5b3 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The address on the internal MDIO bus through which the Lynx PCS is
accessible is actually variable, based on the SGMIIaCR1[MDEV_PORT]
SoC glue layer registers from the SerDes block.

Typically U-Boot configures these registers to good values which
Linux uses blindly (either hardcoded or through the device tree),
but sometimes U-Boot doesn't run.

Now that there exists an API to query the SerDes about the address of a
given protocol converter (the Ethernet PCS), let's use that to warn if
we're not running on a platform that has been prepared in the way that
we expect by previous boot stages.

It is beyond my purposes to dynamically change MDEV_PORT. That would in
fact be quite complicated to do, because for each MAC which goes to a
multi-port PCS like QSGMII, we'd need to know its exact mapping to a
port within that PCS.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>