ILIE-87: media: nxp: neoisp: Rework job scheduling for node groups
commit  90ffcb02871c260715dd576c7072a3c581737b6a from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The scheduling is now grouped as a single method either
preparing the job when node group is ready to be queued
or searching for next group index to be ready.
This change also rework spinlock use to be locked and
unlocked into the same function.

hw_lock attribute is now defined as bool instead
of int variable.

Signed-off-by: Alexi Birlinger <alexi.birlinger@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>