LF-13206: media: ox05b1s: Update controls when format changes
commit  474bdcee46fafa1315a7dcfc738c99982fed9550 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The first run with libcamera does not work because the pixel rate
control value is not updated after set_fmt, it is only updated during
the first streamon. This is causing libcamerasrc negotiation failure
with the downstream caps filter on the fps.
So, update control values to reflect the current format upon
set_fmt.

Signed-off-by: Mirela Rabulea <mirela.rabulea@oss.nxp.com>
Reviewed-by: Alexi Birlinger <alexi.birlinger@nxp.com>
Tested-by: Alexi Birlinger <alexi.birlinger@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>