LF-11880-5 net: pcs: xpcs: reconfig the pcs and phy when the link state is fault
commit  7091f7f8e5414f9b7685d170cba480a10217508d from
https://github.com/nxp-imx/linux-imx lf-6.12.y

We have found that sometimes after the system suspend/resume process,
the connection between the SoC's PHY and the PHY on the board cannot
be established.
As both the SoC's PCS and PHY IPs are provided by the vendor, the
specific reason has not been identified yet.

Therefore, a workaround has been added for this issue in the get_state
API, which involves reconfiguring the PCS and PHY when an abnormal
state of the SoC PHY is detected.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Wei Fang <wei.fang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>