LF-13503-1 net: enetc: use bits to indicate some port capabilities
commit  88ec027e662ee7f8608f7423a87a851959c36d8c from
https://github.com/nxp-imx/linux-imx lf-6.12.y

For some port capabilities, such as WOL and half-duplex, the bool
type was used to indicate whether the port has the capability. In
fact, it is more appropriate to use a single bit to indicate it.
In this way, when adding a new capability, only one more bit needs
to be defined, which will not waste memory.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>