LF-14261-4 net: enetc: add basic support for Pseudo MAC ENETC
commit  ec1531aa40f00b02284ecdd0b4d9147b3cb4f1bb from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

The Pseudo MAC ENETC is an internal port which connects to the CPU port
of the switch. Since Switch CPU/host ENETC is fully integrated with the
switch and does not require a back-to-back MAC, instead a light weight
"pseudo MAC" provides the delineation between switch and ENETC. This
translates to lower power (less logic and memory) and lower delay (as
there is no serialization delay across this link).

Different from the ordinary ENETC (which is used as an external port),
the Pseudo MAC ENETC has different PCIe device ID and less registers,
and some registers are different from ordinary ENETC. So add the basic
support for this Pseudo MAC ENETC. More supports will be added in the
future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Reviewed-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>