LF-14261-13 net: enetc: add RFS table support
commit  0f3d555634eba74054d73c5b938d8948e28ffd08 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

For i.MX94 series, the ENETC supports RFS feature, so add the RFS table
support to the netc-lib driver. Note that the RFS feature of ENETC is
not supported in this patch, it will be implemented in the future.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>