LF-14260-13 net: dsa: netc: add buffer pool support
commit  7946e38e2c4a6cee0544ba56647b711ee1068169 from
https://github.com/nxp-real-time-edge-sw/real-time-edge-linux.git linux_6.12.3-imx943-er1

The buffer pool is a quantity of memory available for buffering a group
of flows (e.g. frames having the same priority, frames received from the
same port), while waiting to be transmitted on a port. The buffer pool
tracks internal memory consumption with upper bound limits and optionally
a non-shared portion when associated with a shared buffer pool. Currently
the shared buffer pool has not been associated with the buffer pool. We
will add the shared buffer pool support in the future. In addition, the
flow control feature of the switch is also implemented through the buffer
pool.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>