LF-13746-6 net: enetc: replace netc-prb-ierb with netc-blk-ctrl driver
commit  29bad0cc99cf19bdbd482529293e364a304ac121 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

There is not much difference between netc-blk-ctrl and netc-prb-ierb,
but the previous naming is not very appropriate, because the driver
also configures NETCMIX block, but the register range of NETCMIX is
obatined through the syscon interface, which may not be appropriate
and not easily accepted by upstream. In the previous patch, we have
divided NETCMIX into two parts, one of which is used by a separate
clock driver. The other part is combined with PRB and IERB to form
netc-blk-ctrl, which is used by the netc-blk-ctrl driver.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>