LF-13446-2 arm64: dts: imx95-19x19-evk: remove regulator-always-on for vbus regulator
commit  443021e368b89d815a84f10d0c59853b979bc352 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

If vbus is left on by the bootloader, vbus will be off when initializing
and then on instantly due to regulator-always-on property. This will
create a glitch condition (~6ms) on vbus line. The usb device may go to
a mess state and usb reset signal cannot reset it too.

To avoid such short glitch, removing the regulator-always-on property
then vbus will be controlled by usb driver and it will have an enough
off period before turning it on.

Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Xu Yang <xu.yang_2@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>