LF-13684-5 media/i2c: max96724: fix "unused value" Coverity issue
commit  528c3008b3507c64d86afd3840784cd02a925aca from
https://github.com/nxp-imx/linux-imx lf-6.12.y

There is a valid issue inside the function where ret is assigned the
returned value of max96724_xlate_streams() in case of error and then we
jump to 'unlock' only to return 0. So, let's use that ret value this
time. Also, initialize ret to 0 to avoid returning wrong value if
pos == -1 in the while loop.

Signed-off-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Reviewed-by: Robert Chiras <robert.chiras@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>