LF-13535-5: media: nxp: add v4l2 subdev driver for parallel csi
commit  12f9d0ee3c4d2f9be3221f025f911ac12005e014 from
https://github.com/nxp-imx/linux-imx lf-6.12.y

The parallel csi is a module which is for parallel sensor's data receiver.
Add a V4L2 subdev driver for it since it's a standalone module and need to
export more controls to userspace. The codes as newly added here.

Signed-off-by: Alice Yuan <alice.yuan@nxp.com>
Reviewed-by: Guoniu Zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>